AI (part 2)
Added a new variable AIObjective that corresponds to the target of the AI (enemy or tokens).

With that variable the AI will know when to attack or when to go look for tokens.

Added an AITimer variable that is used to prevent the AI from acting too fast. This way he can be a bit more natural when playing.