Small changes
Changed Special Tile cooldown sprites and End Turn button sprite.

Added Background image.

Now when attacking instead of being unable to move, the attacker has his moves reduced to 0. This means that they can still move if they choose to use the Extra Moves mechanic.

Reverted a change that made the AI always executing FindAIObjective every turn. This change was originally made because I thought that it was part of a previous bug.